+++ meteor/server/__tests__/api/serviceMessages/serviceMessagesApi.test.ts
@@ -1,12 +1,12 @@
 import { readAllMessages, writeMessage } from '../../../api/serviceMessages/serviceMessagesApi'
 import * as CoreSystemUtil from '../../../coreSystem/collection'
-import { protectString } from '../../../../lib/lib'
+import { protectString } from '../../../lib/tempLib'
 import {
 	Criticality,
 	ExternalServiceMessage,
 	ICoreSystem,
 	ServiceMessage,
-} from '../../../../lib/collections/CoreSystem'
+} from '@sofie-automation/meteor-lib/dist/collections/CoreSystem'
 import { CoreSystem } from '../../../collections'
 import { SupressLogMessages } from '../../../../__mocks__/suppressLogging'