+++ meteor/server/coreSystem/checkDatabaseVersions.ts
@@ -9,10 +9,10 @@ import {
 	parseRange,
 	isPrerelease,
 	parseCoreIntegrationCompatabilityRange,
-} from '../../lib/collections/CoreSystem'
+} from '../systemStatus/semverUtils'
 import { DBShowStyleBase } from '@sofie-automation/corelib/dist/dataModel/ShowStyleBase'
 import { DBStudio } from '@sofie-automation/corelib/dist/dataModel/Studio'
-import { lazyIgnore } from '../../lib/lib'
+import { lazyIgnore } from '../lib/lib'
 import { logger } from '../logging'
 import { CURRENT_SYSTEM_VERSION } from '../migration/currentSystemVersion'
 import { setSystemStatus, removeSystemStatus } from '../systemStatus/systemStatus'