+++ meteor/server/publications/pieceContentStatusUI/rundown/regenerateItems.ts
@@ -6,8 +6,8 @@ import {
 	RundownBaselineAdLibActionId,
 } from '@sofie-automation/corelib/dist/dataModel/Ids'
 import { ReadonlyDeep } from 'type-fest'
-import { UIPieceContentStatus } from '../../../../lib/api/rundownNotifications'
-import { literal, protectString } from '../../../../lib/lib'
+import { UIPieceContentStatus } from '@sofie-automation/meteor-lib/dist/api/rundownNotifications'
+import { literal, protectString } from '../../../lib/tempLib'
 import { CustomPublishCollection } from '../../../lib/customPublication'
 import { ContentCache } from './reactiveContentCache'
 import { wrapTranslatableMessageFromBlueprintsIfNotString } from '@sofie-automation/corelib/dist/TranslatableMessage'