+++ packages/webui/src/client/lib/__tests__/logging.test.ts
@@ -1,0 +1,11 @@
+import { logger } from '../logging'
+
+describe('lib/logger', () => {
+	test('logger', () => {
+		expect(typeof logger.error).toEqual('function')
+		expect(typeof logger.warn).toEqual('function')
+		// expect(typeof logger.help).toEqual('function')
+		expect(typeof logger.info).toEqual('function')
+		expect(typeof logger.debug).toEqual('function')
+	})
+})